+++ pages/my-account/index.tsx
@@ -1,18 +1,25 @@
 import useSWR from 'swr'
+import { useEffect } from 'react';
+import { useRouter } from 'next/router'
 
 import Meta from '@atoms/Meta';
 import RouteGuard from '@atoms/RouteGuard';
-import GetStarted from '@layouts/AccountPage/GetStarted';
 import UserHome from '@layouts/AccountPage/UserHome'
 import Loader from '@atoms/Loader';
 import Layout from '@layouts/AccountPage/Layout';
 import { myFetch } from '@/assets/utilities';
+import { routes } from '@constants';
 
 export default function AccountPage() {
   const { data: profile, isLoading } = useSWR('/api/profile-details', myFetch)
+  const router = useRouter()
 
-  if (isLoading) 
-    return <Loader paddingBlock='10rem' />
+  useEffect(() => {
+    if (!profile?.[0]?.username && !isLoading)
+      router.push('/get-started')
+
+    console.log('shit')
+  }, [profile, isLoading])
 
   return profile?.[0]?.username ?
     <>
@@ -23,10 +30,5 @@ export default function AccountPage() {
         </Layout>
       </RouteGuard>
     </>
-    : <>
-      <Meta title='Get Started | Moviemanic' />
-      <RouteGuard>
-        <GetStarted />
-      </RouteGuard>
-    </>
+    : <Loader paddingBlock='10rem' />
 }
